requirements.txt
discord.py
pillow
python-dotenv
fontTools
